[#7] refactor to fulfil detekt
